requirements.txt
torch
torchvision
torchaudio
numpy
opencv-python
matplotlib
pillow
tqdm
scikit-learn
albumentations
streamlit
onnxruntime
